refactor(engine): move codehash resolution helpers to context::query
The five `pub(crate)` read-only address-resolution helpers
(`bytecode_at`, `resolve_canonical_via_codehash`,
`resolve_analysis_via_codehash`, `resolve_artifact_via_codehash`,
`resolve_recompiled_artifact_via_codehash`) thematically belong with
the other query-style helpers in `context::query` rather than the
lifecycle methods in `context::mod`. Move them, along with the
`Bytes`/`keccak256` imports they need, into a second `impl` block
in `query.rs`. No public API surface changes.
