+++ crates/engine/src/context/mod.rs
@@ -62,7 +62,7 @@ use std::{
     sync::Arc,
 };
 
-use alloy_primitives::{Address, B256, Bytes, TxHash, keccak256};
+use alloy_primitives::{Address, B256, TxHash};
 use edb_common::{
     ForkInfo,
     types::{Trace, parse_callable_abi_entries},
@@ -320,96 +320,6 @@ where
         ));
 
         Ok(())
-    }
-
-    /// Look up the runtime bytecode observed at `address` during Pass 1.
-    ///
-    /// Walks the recorded execution trace for the first entry whose
-    /// `code_address` matches and returns its captured `bytecode`. Returns
-    /// `None` when no trace entry recorded code for that address (e.g. the
-    /// address was never executed, or its frame was elided).
-    ///
-    /// The trace is the canonical Pass-3 source for "what bytes were
-    /// actually executing at this address" — the same bytes the engine
-    /// keccak'd when building [`Self::codehash_to_canonical`] in `prepare`.
-    pub(crate) fn bytecode_at(&self, address: Address) -> Option<Bytes> {
-        self.trace
-            .iter()
-            .find(|entry| entry.code_address == address)
-            .and_then(|entry| entry.bytecode.clone())
-    }
-
-    /// Try to resolve a canonical artifact address for `address` via the
-    /// codehash alias index.
-    ///
-    /// Returns the canonical address whose recompiled instrumented runtime
-    /// hashes to the same digest as the bytecode at `address`. Returns
-    /// `None` when there is no recorded bytecode for `address`, the
-    /// bytecode is empty, or no alias entry exists for that hash.
-    ///
-    /// This mirrors the fallback wired into `HookSnapshotInspector` (see
-    /// `resolve_analysis`) so RPC readers stay consistent with the
-    /// snapshot data the inspector produced for `vm.etch`-aliased
-    /// addresses.
-    pub(crate) fn resolve_canonical_via_codehash(&self, address: Address) -> Option<Address> {
-        let raw_code = self.bytecode_at(address)?;
-        if raw_code.is_empty() {
-            return None;
-        }
-        let hash = keccak256(raw_code.as_ref());
-        let canonical = self.codehash_to_canonical.get(&hash).copied()?;
-        debug!(?address, ?canonical, ?hash, "engine context resolved address via codehash alias");
-        Some(canonical)
-    }
-
-    /// Look up the analysis result for `address`, falling back to the
-    /// codehash-aliased canonical address on miss.
-    ///
-    /// Behaves like `self.analysis_results.get(&address)` for direct hits,
-    /// then for misses retries at the canonical address returned by
-    /// [`Self::resolve_canonical_via_codehash`]. This is the standard
-    /// reader-side fallback for handlers that consume
-    /// `analysis_results` keyed by a user-supplied or trace-derived
-    /// address (notably the etched address recorded on a hook snapshot's
-    /// `bytecode_address` after `vm.etch`).
-    pub(crate) fn resolve_analysis_via_codehash(
-        &self,
-        address: Address,
-    ) -> Option<&AnalysisResult> {
-        self.analysis_results.get(&address).or_else(|| {
-            self.resolve_canonical_via_codehash(address)
-                .and_then(|canonical| self.analysis_results.get(&canonical))
-        })
-    }
-
-    /// Look up the original artifact for `address`, falling back to the
-    /// codehash-aliased canonical address on miss.
-    ///
-    /// Symmetric to [`Self::resolve_analysis_via_codehash`] but for the
-    /// `artifacts` map. Used by code/abi RPC handlers so a `vm.etch`-aliased
-    /// address surfaces the canonical artifact's sources and metadata
-    /// instead of an `INVALID_ADDRESS` or `null`.
-    pub(crate) fn resolve_artifact_via_codehash(&self, address: Address) -> Option<&Artifact> {
-        self.artifacts.get(&address).or_else(|| {
-            self.resolve_canonical_via_codehash(address)
-                .and_then(|canonical| self.artifacts.get(&canonical))
-        })
-    }
-
-    /// Look up the recompiled (instrumented) artifact for `address`,
-    /// falling back to the codehash-aliased canonical address on miss.
-    ///
-    /// Symmetric to [`Self::resolve_artifact_via_codehash`] but for the
-    /// `recompiled_artifacts` map. Backs the ABI handlers, where the
-    /// instrumented artifact's contract ABI is the source of truth.
-    pub(crate) fn resolve_recompiled_artifact_via_codehash(
-        &self,
-        address: Address,
-    ) -> Option<&Artifact> {
-        self.recompiled_artifacts.get(&address).or_else(|| {
-            self.resolve_canonical_via_codehash(address)
-                .and_then(|canonical| self.recompiled_artifacts.get(&canonical))
-        })
     }
 }
